Suport Ascend NPU
Ascend NPU is a series of AI processors.
[hiascend](https://www.hiascend.com/).

This commit make InputArray and OutputArray support a new Mat
structor called AclMat, which has a same interface with Mat and GpuMat.
And also, python bindings generator support AclMat for python interface.

TODO: it's not a very good idea that each new backend adds a new Mat
structor and change gen2.py, it leads to duplication of code and make
python interface inefficient. A better choose is to make new backend Mat
structor inherit of Mat or GpuMat, so that code generator and
Input/Output Array do not need to change. This part is working in
progress.